requirements.txt
chainer==5.0.0b4
cupy==5.0.0b4
Pillow-SIMD==5.1.1.post0
